refactor(cff): extract types module from monolithic cff_subsetter
Move CffDictToken, CffDictScanner, read_u16/u32/i16, otf_checksum, and
encode_cff_int_5byte into a new cff::types module. No behavior changes —
pure file reorganization to prepare for the charstring desubroutinizer.
