+++ oxidize-pdf-core/src/text/fonts/cff/mod.rs
@@ -1,0 +1,6 @@
+//! CFF (Compact Font Format) module.
+//!
+//! This module provides CFF parsing primitives extracted from the monolithic
+//! `cff_subsetter` module, intended to be shared across CFF-related components.
+
+pub mod types;
