+++ oxidize-pdf-core/src/text/fonts/cff_subsetter.rs
@@ -30,152 +30,10 @@
 
 use crate::fonts::cmap_utils::parse_cmap_format_12_filtered;
 use crate::parser::{ParseError, ParseResult};
+use crate::text::fonts::cff::types::{
+    encode_cff_int_5byte, otf_checksum, read_i16, read_u16, read_u32, CffDictScanner, CffDictToken,
+};
 use std::collections::{HashMap, HashSet};
-
-// =============================================================================
-// CFF DICT token scanner
-// =============================================================================
-
-/// A token produced by scanning a CFF DICT byte sequence.
-///
-/// CFF DICTs consist of interleaved operands and operators per CFF spec §4.
-/// The scanner yields one token per call to `next()`.
-#[derive(Debug, PartialEq, Clone)]
-pub enum CffDictToken {
-    /// Integer operand value decoded from the CFF encoding.
-    /// Real numbers (byte 30) emit `Operand(0)` as a placeholder since they
-    /// are not used for any offset calculation.
-    Operand(i32),
-    /// Single-byte operator (bytes 0..=27 except 12, 28, 29, 30).
-    Operator(u8),
-    /// Two-byte escaped operator: first byte was 12, second byte is stored here.
-    EscapedOperator(u8),
-}
-
-/// Iterator over CFF DICT tokens.
-///
-/// Implements the full CFF integer/real operand encoding per CFF spec §4:
-///
-/// | Byte range | Encoding                                      |
-/// |-----------|-----------------------------------------------|
-/// | 32–246    | 1-byte integer: `value = b − 139`            |
-/// | 247–250   | 2-byte positive: `(b0−247)×256 + b1 + 108`  |
-/// | 251–254   | 2-byte negative: `−(b0−251)×256 − b1 − 108` |
-/// | 28        | 2-byte signed big-endian i16                  |
-/// | 29        | 4-byte signed big-endian i32                  |
-/// | 30        | Real number (nibble pairs until 0xF): `Operand(0)` |
-/// | 12        | Escaped operator; reads one more byte         |
-/// | 0–27 (excl. 12, 28, 29, 30) | Single-byte operator    |
-pub struct CffDictScanner<'a> {
-    data: &'a [u8],
-    pos: usize,
-}
-
-impl<'a> CffDictScanner<'a> {
-    /// Create a new scanner over the given CFF DICT data.
-    pub fn new(data: &'a [u8]) -> Self {
-        Self { data, pos: 0 }
-    }
-
-    /// Current byte position within the data slice.
-    pub fn position(&self) -> usize {
-        self.pos
-    }
-}
-
-impl<'a> Iterator for CffDictScanner<'a> {
-    type Item = CffDictToken;
-
-    fn next(&mut self) -> Option<CffDictToken> {
-        loop {
-            if self.pos >= self.data.len() {
-                return None;
-            }
-            let b = self.data[self.pos];
-
-            match b {
-                28 => {
-                    // 2-byte signed integer (big-endian i16)
-                    if self.pos + 2 >= self.data.len() {
-                        return None; // truncated — stop cleanly
-                    }
-                    let v = i16::from_be_bytes([self.data[self.pos + 1], self.data[self.pos + 2]])
-                        as i32;
-                    self.pos += 3;
-                    return Some(CffDictToken::Operand(v));
-                }
-                29 => {
-                    // 4-byte signed integer (big-endian i32)
-                    if self.pos + 4 >= self.data.len() {
-                        return None; // truncated — stop cleanly
-                    }
-                    let v = i32::from_be_bytes([
-                        self.data[self.pos + 1],
-                        self.data[self.pos + 2],
-                        self.data[self.pos + 3],
-                        self.data[self.pos + 4],
-                    ]);
-                    self.pos += 5;
-                    return Some(CffDictToken::Operand(v));
-                }
-                30 => {
-                    // Real number — skip nibble pairs until 0xF terminator
-                    // Emit Operand(0) as a placeholder (real values not used for offsets).
-                    self.pos += 1;
-                    while self.pos < self.data.len() {
-                        let nibble_byte = self.data[self.pos];
-                        self.pos += 1;
-                        if nibble_byte & 0x0F == 0x0F || nibble_byte >> 4 == 0x0F {
-                            break;
-                        }
-                    }
-                    return Some(CffDictToken::Operand(0));
-                }
-                32..=246 => {
-                    // 1-byte integer: value = b − 139
-                    let v = b as i32 - 139;
-                    self.pos += 1;
-                    return Some(CffDictToken::Operand(v));
-                }
-                247..=250 => {
-                    // 2-byte positive integer
-                    if self.pos + 1 >= self.data.len() {
-                        return None; // truncated — stop cleanly
-                    }
-                    let w = self.data[self.pos + 1] as i32;
-                    let v = (b as i32 - 247) * 256 + w + 108;
-                    self.pos += 2;
-                    return Some(CffDictToken::Operand(v));
-                }
-                251..=254 => {
-                    // 2-byte negative integer
-                    if self.pos + 1 >= self.data.len() {
-                        return None; // truncated — stop cleanly
-                    }
-                    let w = self.data[self.pos + 1] as i32;
-                    let v = -(b as i32 - 251) * 256 - w - 108;
-                    self.pos += 2;
-                    return Some(CffDictToken::Operand(v));
-                }
-                12 => {
-                    // Escaped operator: read next byte
-                    self.pos += 1;
-                    if self.pos >= self.data.len() {
-                        return None;
-                    }
-                    let op2 = self.data[self.pos];
-                    self.pos += 1;
-                    return Some(CffDictToken::EscapedOperator(op2));
-                }
-                _ => {
-                    // Single-byte operator (bytes 0–27 excluding 12, 28, 29, 30)
-                    self.pos += 1;
-                    return Some(CffDictToken::Operator(b));
-                }
-            }
-        }
-    }
-}
 
 // =============================================================================
 // Public API
@@ -903,13 +761,6 @@ fn parse_top_dict(data: &[u8]) -> TopDictOffsets {
 // =============================================================================
 
 /// Encode an integer as the 5-byte CFF form (byte 29 + big-endian i32).
-/// Using the fixed-width encoding simplifies two-pass offset calculation:
-/// the byte size of Top DICT is always the same regardless of offset values.
-fn encode_cff_int_5byte(value: i32) -> [u8; 5] {
-    let bytes = value.to_be_bytes();
-    [29, bytes[0], bytes[1], bytes[2], bytes[3]]
-}
-
 /// Rebuild a CID Top DICT, replacing charset (15), CharStrings (17),
 /// FDArray (12 36), and FDSelect (12 37) with new offsets.
 /// All other operators (ROS 12 30, etc.) are preserved verbatim.
@@ -2329,44 +2180,6 @@ fn parse_cmap_format_4(cmap: &[u8], offset: usize) -> ParseResult<HashMap<u32, u
 }
 
 // =============================================================================
-// Utility functions
-// =============================================================================
-
-fn read_u16(data: &[u8], offset: usize) -> ParseResult<u16> {
-    if offset + 2 > data.len() {
-        return Err(ParseError::SyntaxError {
-            position: offset,
-            message: "read_u16: out of bounds".to_string(),
-        });
-    }
-    Ok(u16::from_be_bytes([data[offset], data[offset + 1]]))
-}
-
-fn read_u32(data: &[u8], offset: usize) -> ParseResult<u32> {
-    if offset + 4 > data.len() {
-        return Err(ParseError::SyntaxError {
-            position: offset,
-            message: "read_u32: out of bounds".to_string(),
-        });
-    }
-    Ok(u32::from_be_bytes([
-        data[offset],
-        data[offset + 1],
-        data[offset + 2],
-        data[offset + 3],
-    ]))
-}
-
-fn read_i16(data: &[u8], offset: usize) -> ParseResult<i16> {
-    if offset + 2 > data.len() {
-        return Err(ParseError::SyntaxError {
-            position: offset,
-            message: "read_i16: out of bounds".to_string(),
-        });
-    }
-    Ok(i16::from_be_bytes([data[offset], data[offset + 1]]))
-}
-
 // =============================================================================
 // Type 2 CharString subr-call scanner
 // =============================================================================
@@ -2612,26 +2425,6 @@ fn filter_subr_index(original: &CffIndex, cff: &[u8], used: &HashSet<usize>) -> 
 }
 
 /// Compute an OTF table checksum (sum of big-endian u32 words, padded to 4 bytes).
-fn otf_checksum(data: &[u8]) -> u32 {
-    let mut sum = 0u32;
-    let mut i = 0;
-    while i + 3 < data.len() {
-        sum = sum.wrapping_add(u32::from_be_bytes([
-            data[i],
-            data[i + 1],
-            data[i + 2],
-            data[i + 3],
-        ]));
-        i += 4;
-    }
-    if i < data.len() {
-        let mut last = [0u8; 4];
-        last[..data.len() - i].copy_from_slice(&data[i..]);
-        sum = sum.wrapping_add(u32::from_be_bytes(last));
-    }
-    sum
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
